# Rewrite Array#each in Ruby using Primitive