Same as https://github.com/ruby/ruby/pull/6687, but race condition-free.

## microbenchmark
### Interpreter
Thanks to `Primitive`, the interpreter doesn't slow down.

```
$ benchmark-driver benchmark/loop_each.yml -v --chruby 'before;after'
before: ruby 3.4.0dev (2024-01-20T01:12:07Z master 99d6e2f1ee) [x86_64-linux]
after: ruby 3.4.0dev (2024-01-20T01:36:08Z primitive-array-each 3aaa6de3ec) [x86_64-linux]
Warming up --------------------------------------
           loop_each      1.539 i/s -       2.000 times in 1.299363s (649.68ms/i)
Calculating -------------------------------------
                         before       after
           loop_each      1.551       1.796 i/s -       4.000 times in 2.578413s 2.227167s

Comparison:
                        loop_each
               after:         1.8 i/s
              before:         1.6 i/s - 1.16x  slower
```

### YJIT
Even faster than https://github.com/ruby/ruby/pull/6687.

```
$ benchmark-driver benchmark/loop_each.yml -v --chruby 'before::before --yjit-call-threshold=1;after::after --yjit-call-threshold=1'
before: ruby 3.4.0dev (2024-01-20T01:12:07Z master 99d6e2f1ee) +YJIT [x86_64-linux]
after: ruby 3.4.0dev (2024-01-20T01:36:08Z primitive-array-each 3aaa6de3ec) +YJIT [x86_64-linux]
Warming up --------------------------------------
           loop_each      1.789 i/s -       2.000 times in 1.118068s (559.03ms/i)
Calculating -------------------------------------
                         before       after
           loop_each      1.790      12.986 i/s -       5.000 times in 2.793497s 0.385044s

Comparison:
                        loop_each
               after:        13.0 i/s
              before:         1.8 i/s - 7.26x  slower
```