@maximecb 2024-01-15
Nice. Happy to see this working. Surprised it's so far in YJIT as well!

Ideally, for YJIT, we'd like to be able to avoid doing a function call at all, so we can generate tight inlined code though, so we might need a specialized instruction instead of a cexpr? What do you think?